break PathEntry classes into separate files, a bit of cleanup
